I can't believe it. Something is coming up

+++ src/utils/ppu_utils.c
@@ -25,7 +25,7 @@ void encode_as_tiles(byte *mem_addr, byte number_tiles, struct tile *tiles) {
 	}
 }
 
-struct tile nametable_tile(uint tile_id) {
+struct tile nametable_tile(byte tile_id) {
 	uint base_tile_addr = bit_test(rmem_b(PPUCTRL), PPUCTRL_S) ? 0x0000 : 0x1000;
 	base_tile_addr += tile_id * 16;
